+++ packages/app-check/src/internal-api.test.ts
@@ -111,7 +111,7 @@ describe('internal api', () => {
 
       expect(reCAPTCHASpy).to.be.called;
 
-      expect(exchangeTokenStub.args[0][0].body['recaptcha_token']).to.equal(
+      expect(exchangeTokenStub.args[0][0].body['recaptcha_v3_token']).to.equal(
         fakeRecaptchaToken
       );
       expect(token).to.deep.equal({ token: fakeRecaptchaAppCheckToken.token });