+++ packages/auth/src/model/public_types.ts
@@ -101,6 +101,7 @@ export interface ParsedToken {
   'firebase'?: {
     'sign_in_provider'?: string;
     'sign_in_second_factor'?: string;
+    'identities'?: Record<string, string>;
   };
   /** Map of any additional custom claims. */
   [key: string]: string | object | undefined;
@@ -246,13 +247,28 @@ export interface Auth {
    * To keep the old behavior, see {@link Auth.onIdTokenChanged}.
    *
    * @param nextOrObserver - callback triggered on change.
-   * @param error - callback triggered on error.
-   * @param completed - callback triggered when observer is removed.
+   * @param error - Deprecated. This callback is never triggered. Errors
+   * on signing in/out can be caught in promises returned from
+   * sign-in/sign-out functions.
+   * @param completed - Deprecated. This callback is never triggered.
    */
   onAuthStateChanged(
     nextOrObserver: NextOrObserver<User | null>,
     error?: ErrorFn,
     completed?: CompleteFn
+  ): Unsubscribe;
+  /**
+   * Adds a blocking callback that runs before an auth state change
+   * sets a new user.
+   *
+   * @param callback - callback triggered before new user value is set.
+   *   If this throws, it blocks the user from being set.
+   * @param onAbort - callback triggered if a later `beforeAuthStateChanged()`
+   *   callback throws, allowing you to undo any side effects.
+   */
+  beforeAuthStateChanged(
+    callback: (user: User | null) => void | Promise<void>,
+    onAbort?: () => void,
   ): Unsubscribe;
   /**
    * Adds an observer for changes to the signed-in user's ID token.
@@ -261,8 +277,10 @@ export interface Auth {
    * This includes sign-in, sign-out, and token refresh events.
    *
    * @param nextOrObserver - callback triggered on change.
-   * @param error - callback triggered on error.
-   * @param completed - callback triggered when observer is removed.
+   * @param error - Deprecated. This callback is never triggered. Errors
+   * on signing in/out can be caught in promises returned from
+   * sign-in/sign-out functions.
+   * @param completed - Deprecated. This callback is never triggered.
    */
   onIdTokenChanged(
     nextOrObserver: NextOrObserver<User | null>,