+++ packages/auth/test/integration/flows/custom.local.test.ts
@@ -39,6 +39,7 @@ import {
   getTestInstance,
   randomEmail
 } from '../../helpers/integration/helpers';
+import { generateMiddlewareTests } from './middleware_test_generator';
 
 use(chaiAsPromised);
 
@@ -224,5 +225,9 @@ describe('Integration test: custom auth', () => {
         'auth/email-already-in-use'
       );
     });
+  });
+
+  generateMiddlewareTests(() => auth, () => {
+    return signInWithCustomToken(auth, customToken);
   });
 });