+++ packages/database-compat/src/api/Database.ts
@@ -19,6 +19,8 @@
 import { FirebaseApp } from '@firebase/app-types';
 import { FirebaseService } from '@firebase/app-types/private';
 import {
+  forceLongPolling,
+  forceWebSockets,
   goOnline,
   connectDatabaseEmulator,
   goOffline,
@@ -51,7 +53,9 @@ export class Database implements FirebaseService, Compat<ModularDatabase> {
   constructor(readonly _delegate: ModularDatabase, readonly app: FirebaseApp) {}
 
   INTERNAL = {
-    delete: () => this._delegate._delete()
+    delete: () => this._delegate._delete(),
+    forceWebSockets,
+    forceLongPolling
   };
 
   /**