+++ packages/firebase/compat/index.d.ts
@@ -5215,6 +5215,9 @@ declare namespace firebase.analytics {
 
     /**
      * Use gtag 'config' command to set 'screen_name'.
+     *
+     * @deprecated Use {@link logEvent} with `eventName` as 'screen_view' and add relevant `eventParams`.
+     * See {@link https://firebase.google.com/docs/analytics/screenviews | Track Screenviews}.
      */
     setCurrentScreen(
       screenName: string,
@@ -6018,6 +6021,20 @@ declare namespace firebase.database {
      * ```
      */
     app: firebase.app.App;
+    /**
+     * Additional methods for debugging and special cases.
+     *
+     */
+    INTERNAL: {
+      /**
+       * Force the use of WebSockets instead of long polling.
+       */
+      forceWebSockets: () => void;
+      /**
+       * Force the use of long polling instead of WebSockets. This will be ignored if the WebSocket protocol is used in `databaseURL`.
+       */
+      forceLongPolling: () => void;
+    };
     /**
      * Modify this instance to communicate with the Realtime Database emulator.
      *