+++ packages/firestore/src/core/query.ts
@@ -19,7 +19,6 @@ import { compareDocumentsByField, Document } from '../model/document';
 import { DocumentKey } from '../model/document_key';
 import { FieldPath, ResourcePath } from '../model/path';
 import { debugAssert, debugCast, fail } from '../util/assert';
-import { isNullOrUndefined } from '../util/types';
 
 import {
   Bound,
@@ -155,7 +154,7 @@ export function asCollectionQueryAtPath(
  * Returns true if this query does not specify any query constraints that
  * could remove results.
  */
-export function matchesAllDocuments(query: Query): boolean {
+export function queryMatchesAllDocuments(query: Query): boolean {
   return (
     query.filters.length === 0 &&
     query.limit === null &&
@@ -165,14 +164,6 @@ export function matchesAllDocuments(query: Query): boolean {
       (query.explicitOrderBy.length === 1 &&
         query.explicitOrderBy[0].field.isKeyField()))
   );
-}
-
-export function hasLimitToFirst(query: Query): boolean {
-  return !isNullOrUndefined(query.limit) && query.limitType === LimitType.First;
-}
-
-export function hasLimitToLast(query: Query): boolean {
-  return !isNullOrUndefined(query.limit) && query.limitType === LimitType.Last;
 }
 
 export function getFirstOrderByField(query: Query): FieldPath | null {
@@ -324,10 +315,10 @@ export function queryToTarget(query: Query): Target {
 
       // We need to swap the cursors to match the now-flipped query ordering.
       const startAt = queryImpl.endAt
-        ? new Bound(queryImpl.endAt.position, !queryImpl.endAt.inclusive)
+        ? new Bound(queryImpl.endAt.position, queryImpl.endAt.inclusive)
         : null;
       const endAt = queryImpl.startAt
-        ? new Bound(queryImpl.startAt.position, !queryImpl.startAt.inclusive)
+        ? new Bound(queryImpl.startAt.position, queryImpl.startAt.inclusive)
         : null;
 
       // Now return as a LimitType.First query.
@@ -393,7 +384,7 @@ export function queryWithAddedOrderBy(query: Query, orderBy: OrderBy): Query {
 
 export function queryWithLimit(
   query: Query,
-  limit: number,
+  limit: number | null,
   limitType: LimitType
 ): Query {
   return new QueryImpl(