+++ packages/functions/src/api.ts
@@ -24,7 +24,8 @@ import {
   FunctionsService,
   DEFAULT_REGION,
   connectFunctionsEmulator as _connectFunctionsEmulator,
-  httpsCallable as _httpsCallable
+  httpsCallable as _httpsCallable,
+  httpsCallableFromURL as _httpsCallableFromURL
 } from './service';
 import { getModularInstance } from '@firebase/util';
 
@@ -87,6 +88,26 @@ export function httpsCallable<RequestData = unknown, ResponseData = unknown>(
   return _httpsCallable<RequestData, ResponseData>(
     getModularInstance<FunctionsService>(functionsInstance as FunctionsService),
     name,
+    options
+  );
+}
+
+/**
+ * Returns a reference to the callable HTTPS trigger with the specified url.
+ * @param url - The url of the trigger.
+ * @public
+ */
+export function httpsCallableFromURL<
+  RequestData = unknown,
+  ResponseData = unknown
+>(
+  functionsInstance: Functions,
+  url: string,
+  options?: HttpsCallableOptions
+): HttpsCallable<RequestData, ResponseData> {
+  return _httpsCallableFromURL<RequestData, ResponseData>(
+    getModularInstance<FunctionsService>(functionsInstance as FunctionsService),
+    url,
     options
   );
 }