+++ packages/storage-compat/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @firebase/storage-compat
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [[`9c5c9c36d`](https://github.com/firebase/firebase-js-sdk/commit/9c5c9c36da80b98b73cfd60ef2e2965087e9f801)]:
+  - @firebase/util@1.6.0
+  - @firebase/component@0.5.14
+  - @firebase/storage@0.9.6
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [[`e9e5f6b3c`](https://github.com/firebase/firebase-js-sdk/commit/e9e5f6b3ca9d61323b22f87986d9959f5297ec59)]:
+  - @firebase/util@1.5.2
+  - @firebase/component@0.5.13
+  - @firebase/storage@0.9.5
+
 ## 0.1.12
 
 ### Patch Changes